Cleaned up code reducing the number of files. Fixed minor typos in comments
