+++ main.py
@@ -1,3 +1,4 @@
+import runner
 import retrieval
 import commentor
 
@@ -5,6 +6,6 @@ IDs = retrieval.postRetrieve()
 print(IDs)
 
 for i in range(len(IDs)):
-	if commentor.checkComments(IDs[i]) == True:
-		#commentor.post(IDs[i])
-		print('commenting')
+    if commentor.checkComments(IDs[i]) == True:
+        #commentor.post(IDs[i])
+        print('commenting')
